+++ src/datarobot_genai/drmcp/core/dr_mcp_server.py
@@ -27,6 +27,9 @@ from fastmcp.resources import Resource
 from fastmcp.tools import Tool
 from starlette.middleware import Middleware
 
+from datarobot_genai.drmcp.core.feature_flags import FeatureFlag
+from datarobot_genai.drmcp.core.lineage.enums import LRSEnvVarIsNotSetError
+from datarobot_genai.drmcp.core.lineage.manager import LineageManager
 from datarobot_genai.drtools.core.auth import initialize_oauth_middleware
 from datarobot_genai.drtools.core.credentials import get_credentials
 
@@ -240,6 +243,16 @@ class DataRobotMCPServer:
             if self._config.mcp_server_register_dynamic_prompts_on_startup:
                 self._logger.info("Registering dynamic prompts from prompt management...")
                 asyncio.run(register_prompts_from_datarobot_prompt_management())
+
+            if FeatureFlag.is_mcp_tools_gallery_support_enabled():
+                try:
+                    linear_manager = LineageManager(self._mcp)
+                    asyncio.run(linear_manager.sync_mcp_tools())
+                    asyncio.run(linear_manager.sync_mcp_prompts())
+                    self._logger.info("Sync with MCP lineage")
+                except LRSEnvVarIsNotSetError as error:
+                    error_message = f"MCP item metadata is not sync. {str(error)}"
+                    self._logger.warning(error_message)
 
             # Execute pre-server start actions
             asyncio.run(self._lifecycle.pre_server_start(self._mcp))